+++ src/main/java/org/jsoup/select/NodeVisitor.java
@@ -4,11 +4,11 @@ import org.jsoup.nodes.Element;
 import org.jsoup.nodes.Node;
 
 /**
- Node visitor interface. Provide an implementing class to {@link NodeTraversor} or to {@link Node#traverse(NodeVisitor)}
- to iterate through nodes.
+ Node visitor interface, used to walk the DOM, and visit each Node. Execute via {@link #traverse(Node)} or
+ {@link Node#traverse(NodeVisitor)}. The traversal is depth-first.
  <p>
- This interface provides two methods, {@link #head} and {@link #tail}. The head method is called when the node is first
- seen, and the tail method when all of the node's children have been visited. As an example, {@code head} can be used to
+ This interface provides two methods, {@link #head} and {@link #tail}. The head method is called when a node is first
+ seen, and the tail method when all that node's children have been visited. As an example, {@code head} can be used to
  emit a start tag for a node, and {@code tail} to create the end tag. The {@code tail} method defaults to a no-op, so
  the {@code head} method is the {@link FunctionalInterface}.
  </p>
@@ -49,5 +49,14 @@ public interface NodeVisitor {
      */
     default void tail(Node node, int depth) {
         // no-op by default, to allow just specifying the head() method
+    }
+
+    /**
+     Run a depth-first traverse of the root and all of its descendants.
+     @param root the initial node point to traverse.
+     @since 1.21.1
+     */
+    default void traverse(Node root) {
+        NodeTraversor.traverse(this, root);
     }
 }